move working directory context manager to myna.core.utils
