+++ src/myna/core/utils/__init__.py
@@ -13,3 +13,4 @@ from .conversion import str_to_list
 from .get_argparse_defaults import get_script_call_with_defaults
 from .downsample_to_image import *
 from .get_adjacent_layers import *
+from .filesystem import *
